backend/requirements.txt
# ── Core LLM ──────────────────────────────────────────────────────────────────
groq>=0.9.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# ── Memory ────────────────────────────────────────────────────────────────────
mem0ai>=0.1.0

# ── Web Scraping & Search ─────────────────────────────────────────────────────
firecrawl-py>=1.0.0
requests>=2.31.0

# ── PDF Parsing ───────────────────────────────────────────────────────────────
pypdf>=4.0.0
PyPDF2>=3.0.0

# ── Backend (FastAPI) ─────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
